Edit and delete function in inventory and Chef side function with inventory

+++ Spring Boot - Back End/src/main/java/com/kingsman/Kingsman/model/InventoryItem.java
@@ -15,7 +15,7 @@ public class InventoryItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String itemName;
-    private int quantity;
+    private Long quantity;
     private String vendorId;
     @Column(name = "dateAndTime")
     private LocalDateTime dateTime;
